fix: some error

+++ recruitment-common/src/main/java/com/shuyi/recruitment/common/dto/tencent/PostDTO.java
@@ -10,7 +10,7 @@ public class PostDTO {
     private String postId;
 
     @JsonProperty("RecruitPostId")
-    private Integer recruitPostID;
+    private Long recruitPostID;
 
     @JsonProperty("RecruitPostName")
     private String recruitPostName;
